+++ src/Python/verification/verify_calculations.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 import generate_data
-from validation_files_util import read_reference_csvs
+from verification_files_util import read_reference_csvs
 
 impact_dataframe, angles_dataframe, dispersion_dataframe, post_penetration_dataframe, trajectory_dataframe = generate_data.get_dataframes()
